HW_8 corrected

+++ src/main/webapp/resources/js/topjava.common.js
@@ -1,13 +1,93 @@
-var form;
+// var form;
+//
+// function makeEditable(datatableApi) {
+//     ctx.datatableApi = datatableApi;
+//     form = $('#detailsForm');
+//     $(".delete").click(function () {
+//         if (confirm('Are you sure?')) {
+//             deleteRow($(this).closest('tr').attr("id"));
+//         }
+//     });
+//
+//     $(document).ajaxError(function (event, jqXHR, options, jsExc) {
+//         failNoty(jqXHR);
+//     });
+//
+//     // solve problem with cache in IE: https://stackoverflow.com/a/4303862/548473
+//     $.ajaxSetup({cache: false});
+// }
+//
+// function add() {
+//     form.find(":input").val("");
+//     $("#editRow").modal();
+// }
+//
+// function deleteRow(id) {
+//     $.ajax({
+//         url: ctx.ajaxUrl + id,
+//         type: "DELETE"
+//     }).done(function () {
+//         updateTable();
+//         successNoty("Deleted");
+//     });
+// }
+//
+// function updateTable() {
+//     $.get(ctx.ajaxUrl, function (data) {
+//         ctx.datatableApi.clear().rows.add(data).draw();
+//     });
+// }
+//
+// function save() {
+//     $.ajax({
+//         type: "POST",
+//         url: ctx.ajaxUrl,
+//         data: form.serialize()
+//     }).done(function () {
+//         $("#editRow").modal("hide");
+//         ctx.updateTable;
+//         successNoty("Saved");
+//     });
+// }
+//
+// let failedNote;
+//
+// function closeNoty() {
+//     if (failedNote) {
+//         failedNote.close();
+//         failedNote = undefined;
+//     }
+// }
+//
+// function successNoty(text) {
+//     closeNoty();
+//     new Noty({
+//         text: "<span class='fa fa-lg fa-check'></span> &nbsp;" + text,
+//         type: 'success',
+//         layout: "bottomRight",
+//         timeout: 1000
+//     }).show();
+// }
+//
+// function failNoty(jqXHR) {
+//     closeNoty();
+//     failedNote = new Noty({
+//         text: "<span class='fa fa-lg fa-exclamation-circle'></span> &nbsp;Error status: " + jqXHR.status,
+//         type: "error",
+//         layout: "bottomRight"
+//     });
+//     failedNote.show()
+// }
+//
+// function updateTableByData(data) {
+//     ctx.datatableApi.clear().rows.add(data).draw();
+// }
+
+let form;
 
 function makeEditable(datatableApi) {
     ctx.datatableApi = datatableApi;
     form = $('#detailsForm');
-    $(".delete").click(function () {
-        if (confirm('Are you sure?')) {
-            deleteRow($(this).closest('tr').attr("id"));
-        }
-    });
 
     $(document).ajaxError(function (event, jqXHR, options, jsExc) {
         failNoty(jqXHR);
@@ -18,24 +98,36 @@ function makeEditable(datatableApi) {
 }
 
 function add() {
+    $("#modalTitle").html(i18n["addTitle"]);
     form.find(":input").val("");
     $("#editRow").modal();
 }
 
-function deleteRow(id) {
-    $.ajax({
-        url: ctx.ajaxUrl + id,
-        type: "DELETE"
-    }).done(function () {
-        updateTable();
-        successNoty("Deleted");
+function updateRow(id) {
+    form.find(":input").val("");
+    $("#modalTitle").html(i18n["editTitle"]);
+    $.get(ctx.ajaxUrl + id, function (data) {
+        $.each(data, function (key, value) {
+            form.find("input[name='" + key + "']").val(value);
+        });
+        $('#editRow').modal();
     });
 }
 
-function updateTable() {
-    $.get(ctx.ajaxUrl, function (data) {
-        ctx.datatableApi.clear().rows.add(data).draw();
-    });
+function deleteRow(id) {
+    if (confirm(i18n['common.confirm'])) {
+        $.ajax({
+            url: ctx.ajaxUrl + id,
+            type: "DELETE"
+        }).done(function () {
+            ctx.updateTable();
+            successNoty("common.deleted");
+        });
+    }
+}
+
+function updateTableByData(data) {
+    ctx.datatableApi.clear().rows.add(data).draw();
 }
 
 function save() {
@@ -45,8 +137,8 @@ function save() {
         data: form.serialize()
     }).done(function () {
         $("#editRow").modal("hide");
-        ctx.updateTable;
-        successNoty("Saved");
+        ctx.updateTable();
+        successNoty("common.saved");
     });
 }
 
@@ -59,26 +151,34 @@ function closeNoty() {
     }
 }
 
-function successNoty(text) {
+function successNoty(key) {
     closeNoty();
     new Noty({
-        text: "<span class='fa fa-lg fa-check'></span> &nbsp;" + text,
+        text: "<span class='fa fa-lg fa-check'></span> &nbsp;" + i18n[key],
         type: 'success',
         layout: "bottomRight",
         timeout: 1000
     }).show();
 }
 
+function renderEditBtn(data, type, row) {
+    if (type === "display") {
+        return "<a onclick='updateRow(" + row.id + ");'><span class='fa fa-pencil'></span></a>";
+    }
+}
+
+function renderDeleteBtn(data, type, row) {
+    if (type === "display") {
+        return "<a onclick='deleteRow(" + row.id + ");'><span class='fa fa-remove'></span></a>";
+    }
+}
+
 function failNoty(jqXHR) {
     closeNoty();
     failedNote = new Noty({
-        text: "<span class='fa fa-lg fa-exclamation-circle'></span> &nbsp;Error status: " + jqXHR.status,
+        text: "<span class='fa fa-lg fa-exclamation-circle'></span> &nbsp;" + i18n["common.errorStatus"] + ": " + jqXHR.status + (jqXHR.responseJSON ? "<br>" + jqXHR.responseJSON : ""),
         type: "error",
         layout: "bottomRight"
     });
     failedNote.show()
-}
-
-function updateTableByData(data) {
-    ctx.datatableApi.clear().rows.add(data).draw();
 }
